feat(nvim): add directories for specific plugins and config
