+++ README.md
@@ -1,4 +1,5 @@
 # Dotfiles with Ansible
+
 <p align="center">
     <a href="https://github.com/elidholm/dotfiles/actions/workflows/linting.yml"><img align="center" src="https://github.com/elidholm/dotfiles/actions/workflows/linting.yml/badge.svg" alt="github actions"></a>
     <a href="https://github.com/zricethezav/gitleaks-action"><img align="center" src="https://img.shields.io/badge/protected%20by-gitleaks-blue" alt="gitleaks badge"></a>
@@ -40,7 +41,7 @@ Verify your system has all latest packages installed before running the playbook
 sudo apt-get update && sudo apt-get upgrade -y
 ```
 
-> NOTE: This will take some time.
+> [!NOTE] This will take some time.
 
 ## Setup
 
@@ -59,7 +60,6 @@ Below is a list of all available values. Not all are required but incorrect valu
 | system_host           | dict (see [System Hosts](#system-hosts) below)     | no       |
 | bash_public           | dict (see [Environment](#environment) below)       | no       |
 | bash_private          | dict (see [Environment](#environment) below)       | no       |
-
 
 #### Environment
 
@@ -110,14 +110,13 @@ ssh_key:
     3437
 ```
 
-> NOTE: All ssh keys will be stored at `$HOME/.ssh/<filename>`.
+> [!NOTE] All ssh keys will be stored at `$HOME/.ssh/<filename>`.
 
 ### System Hosts
 
 Manage `/etc/hosts` by setting the `system_host` value in `all.yml`.
 
 ```yaml
-
 ---
 system_host:
   127.0.0.1: localhost
@@ -181,11 +180,11 @@ vim ~/.ansible-vault/vault.secret
 To then encrypt values with your vault password use the following:
 
 ```bash
-$ ansible-vault encrypt_string --vault-password-file $HOME/.ansible-vault/vault.secret "mynewsecret" --name "MY_SECRET_VAR"
-$ cat myfile.conf | ansible-vault encrypt_string --vault-password-file $HOME/.ansible-vault/vault.secret --stdin-name "myfile"
+ansible-vault encrypt_string --vault-password-file $HOME/.ansible-vault/vault.secret "mynewsecret" --name "MY_SECRET_VAR"
+cat myfile.conf | ansible-vault encrypt_string --vault-password-file $HOME/.ansible-vault/vault.secret --stdin-name "myfile"
 ```
 
-> NOTE: This file will automatically be detected by the playbook when running `dotfiles` command to decrypt values. Read more on Ansible Vault [here](https://docs.ansible.com/ansible/latest/user_guide/vault.html).
+> [!NOTE] This file will automatically be detected by the playbook when running `dotfiles` command to decrypt values. Read more on [Ansible Vault](https://docs.ansible.com/ansible/latest/user_guide/vault.html).
 
 ## Usage
 
@@ -195,13 +194,14 @@ This playbook includes a custom shell script located at `bin/dotfiles`. This scr
 
 This shell script is also used to initialize your environment after installing `Ubuntu` and performing a full system upgrade as mentioned above.
 
-> NOTE: You must follow required steps before running this command or things may become unusable until fixed.
+> [!NOTE] You must follow required steps before running this command or things may become unusable until fixed.
 
 ```bash
 bash -c "$(curl -fsSL https://raw.githubusercontent.com/elidholm/dotfiles/master/bin/dotfiles)"
 ```
 
 If you want to run only a specific role, you can specify the following bash command:
+
 ```bash
 curl -fsSL https://raw.githubusercontent.com/elidholm/dotfiles/master/bin/dotfiles | bash -s -- --tags comma,seperated,tags
 ```
@@ -229,6 +229,7 @@ This `dotfiles` command is available to you after the first use of this repo, as
 Any flags or arguments you pass to the `dotfiles` command are passed as-is to the `ansible-playbook` command.
 
 For Example: Running the tmux tag with verbosity
+
 ```bash
 dotfiles -t tmux -vvv
 ```
