+++ roles/neovim/files/lua/extra/config/autocmd.lua
@@ -1,0 +1,17 @@
+vim.api.nvim_create_autocmd("TextYankPost", {
+	desc = "Highlight when yanking (copying) text",
+	group = vim.api.nvim_create_augroup("highlight-yank", { clear = true }),
+	pattern = "*",
+	callback = function()
+		vim.hl.on_yank({
+			higroup = "IncSearch",
+			timeout = 40,
+		})
+	end,
+})
+
+vim.api.nvim_create_autocmd({ "BufWritePre" }, {
+	group = vim.api.nvim_create_augroup("conf", {}),
+	pattern = "*",
+	command = [[%s/\s\+$//e]],
+})
