+++ roles/neovim/files/lua/extra/config/work.lua
@@ -1,0 +1,5 @@
+vim.filetype.add({
+	filename = {
+		["Taskfile"] = "yaml",
+	},
+})
