+++ roles/neovim/files/lua/extra/plugins/colorscheme.lua
@@ -1,0 +1,3 @@
+vim.cmd.colorscheme("catppuccin")
+vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
+vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
