+++ roles/neovim/files/lua/extra/plugins/conform.lua
@@ -1,0 +1,33 @@
+require("conform").setup({
+	notify_on_error = false,
+	format_on_save = function(bufnr)
+		local enabled_filetypes = {
+			lua = true,
+			python = true,
+			sh = true,
+		}
+		if enabled_filetypes[vim.bo[bufnr].filetype] then
+			return { timeout_ms = 500 }
+		else
+			return nil
+		end
+	end,
+	default_format_opts = {
+		lsp_format = "fallback",
+	},
+	formatters_by_ft = {
+		lua = { "stylua" },
+		rust = { "rustfmt" },
+		python = { "isort", "black" },
+		sh = { "shfmt" },
+	},
+	formatters = {
+		shfmt = {
+			append_args = { "-i", "2" },
+		},
+	},
+})
+
+vim.keymap.set({ "n", "v" }, "<leader>f", function()
+	require("conform").format({ async = true })
+end, { desc = "[F]ormat buffer" })
