+++ roles/neovim/files/lua/extra/plugins/harpoon.lua
@@ -1,0 +1,3 @@
+require("harpoon").setup({})
+vim.keymap.set("n", "<leader>a", require("harpoon.mark").add_file)
+vim.keymap.set("n", "<C-e>", require("harpoon.ui").toggle_quick_menu)
