+++ roles/neovim/files/lua/extra/plugins/treesitter.lua
@@ -1,0 +1,50 @@
+local parsers = {
+	"lua",
+	"vim",
+	"python",
+	"bash",
+	"css",
+	"csv",
+	"dockerfile",
+	"groovy",
+	"html",
+	"json",
+	"rust",
+	"tsv",
+	"xml",
+	"yaml",
+}
+require("nvim-treesitter").install(parsers)
+
+---@param buf integer
+---@param language string
+local function treesitter_try_attach(buf, language)
+	if not vim.treesitter.language.add(language) then
+		return
+	end
+	vim.treesitter.start(buf, language)
+end
+
+local available_parsers = require("nvim-treesitter").get_available()
+vim.api.nvim_create_autocmd("FileType", {
+	callback = function(args)
+		local buf, filetype = args.buf, args.match
+
+		local language = vim.treesitter.language.get_lang(filetype)
+		if not language then
+			return
+		end
+
+		local installed_parsers = require("nvim-treesitter").get_installed("parsers")
+
+		if vim.tbl_contains(installed_parsers, language) then
+			treesitter_try_attach(buf, language)
+		elseif vim.tbl_contains(available_parsers, language) then
+			require("nvim-treesitter").install(language):await(function()
+				treesitter_try_attach(buf, language)
+			end)
+		else
+			treesitter_try_attach(buf, language)
+		end
+	end,
+})
